+++ services/web/server/tests/unit/isolated/test_exceptions_handlers_base.py
@@ -10,7 +10,7 @@ import pytest
 from aiohttp import web
 from aiohttp.test_utils import make_mocked_request
 from simcore_service_webserver.errors import WebServerBaseError
-from simcore_service_webserver.exceptions_handlers_base import (
+from simcore_service_webserver.exception_handling_base import (
     AiohttpExceptionHandler,
     ExceptionHandlingContextManager,
     _sort_exceptions_by_specificity,
