+++ services/web/server/tests/unit/isolated/test_exceptions_handlers_http_error_map.py
@@ -14,11 +14,11 @@ from aiohttp.test_utils import make_mocked_request
 from servicelib.aiohttp import status
 from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
 from simcore_service_webserver.errors import WebServerBaseError
-from simcore_service_webserver.exceptions_handlers_base import (
+from simcore_service_webserver.exception_handling_base import (
     ExceptionHandlingContextManager,
     exception_handling_decorator,
 )
-from simcore_service_webserver.exceptions_handlers_http_error_map import (
+from simcore_service_webserver.exception_handling_http import (
     ExceptionToHttpErrorMap,
     HttpErrorInfo,
     create_exception_handler_from_http_info,
